Avoid dependency cycle
